test-hygiene: 修复硬编码凭据 + 升级 opentelemetry_sdk + 覆盖率达 95%+
- 安全 MEDIUM: 提取 test_helpers::create_test_db_pool，删除 19 处硬编码 DB 凭据模式
  - 删除 config/test_server.toml (含明文凭据)
  - 新增 config/test_server.example.toml 凭据模板
  - .gitignore 加入 config/test_server.toml
- 安全 LOW: validate_savepoint_name 收紧为 ASCII-only
  - 旧 is_alphanumeric 在 SQL_ASCII/LATIN1 编码下会运行时失败
  - 4 个相关测试断言更新为 rejected
- LOW (CVE-2026-48504): opentelemetry_sdk 0.23.0 → vendor stub crate 隔离
  - 新增 vendor/opentelemetry-jaeger-stub 本地 stub crate
  - 发布前需改回远程版本
- 架构 H1: 4 处 make_db_pool 集中化为 create_test_db_pool
  - sdk/tests.rs / crawl_handler.rs / scrape_handler.rs / distributed_rate_limit_middleware.rs
- 架构 H2: docker-compose.test.yml test-db 服务参数化
  - POSTGRES_USER/PASSWORD/DB 通过环境变量传入
- 架构 M1+M2: 引入 test-mocks feature 统一 mocks 模块可见性
  - production binary 默认不编译 mocks (避免 vtable 残留)
  - integration test 通过 --features test-mocks 启用
- 架构 L1-L5: clippy 修复
  - L1: tests.rs 注释与代码同步
  - L2: 删除误加的 too_many_arguments allow
  - L3: 移除 dead_code allow 暴露真实 dead code
  - L4: 恢复 unwrap_or 测试语义 + 函数级 unnecessary_literal_unwrap allow
  - L5: FindByIdMode 枚举变体重命名 (ReturnTask→Task 等)
- 覆盖率: no-default 95.86% / standard 95.24% / full 95.04% Line
- 单测: 5573 passed / 0 failed
- clippy: 0 errors / 0 warnings
- fmt: 无 diff

+++ src/engines/router.rs
@@ -2054,7 +2054,7 @@ mod tests {
                     screenshot: None,
                     content_type: "text/html".to_string(),
                     headers: HashMap::new(),
-                    response_time_ms: self.delay_ms as u64,
+                    response_time_ms: self.delay_ms,
                 })
             }
             fn support_score(&self, _request: &InternalScrapeRequest) -> u8 {
